+++ src/nofunc-htbl/hashtbl0.mli
@@ -91,7 +91,7 @@ val copy : ('a, 'b) t -> ('a, 'b) t
     (Same behavior as with association lists.)
 
     If you desire the classic behavior of replacing elements,
-    see {!replace}. *)
+    see {!set}. *)
 val add : ('a, 'b) t -> key:'a -> data:'b -> unit
 
 (** [find tbl key] returns the current binding of [key] in [tbl], or [None]
@@ -118,15 +118,15 @@ val remove : ('a, 'b) t -> 'a -> unit
 (** Same as {!remove} but returns the previous binding, if any. *)
 val find_and_remove : ('a, 'b) t -> 'a -> 'b option
 
-(** [replace tbl ~key ~data] replaces the current binding of [key] in [tbl] by a
+(** [set tbl ~key ~data] replaces the current binding of [key] in [tbl] by a
     binding of [key] to [data]. If [key] is unbound in [tbl], a binding of [key]
     to [data] is added to [tbl].
 
     This is functionally equivalent to {!remove}[ tbl key] followed by
     {!add}[ tbl ~key ~data]. *)
-val replace : ('a, 'b) t -> key:'a -> data:'b -> unit
+val set : ('a, 'b) t -> key:'a -> data:'b -> unit
 
-(** Same as {!replace} but returns the previous binding, if any. *)
+(** Same as {!set} but returns the previous binding, if any. *)
 val find_and_replace : ('a, 'b) t -> key:'a -> data:'b -> 'b option
 
 (** [iter tbl ~f] applies [f] to all bindings in table [tbl]. [key] and [data]
@@ -202,8 +202,8 @@ val to_seq_values : (_, 'b) t -> 'b Seq.t
 (** Add the given bindings to the table, using {!add}. *)
 val add_seq : ('a, 'b) t -> ('a * 'b) Seq.t -> unit
 
-(** Add the given bindings to the table, using {!replace}. *)
-val replace_seq : ('a, 'b) t -> ('a * 'b) Seq.t -> unit
+(** Add the given bindings to the table, using {!set}. *)
+val set_seq : ('a, 'b) t -> ('a * 'b) Seq.t -> unit
 
 (** [create_seeded (module Key) n] creates a new, empty hash table, with initial
     size greater or equal to the suggested size [n]. For best results, [n]
@@ -312,3 +312,17 @@ val dyn_of_m__t
   -> ('data -> Dyn.t)
   -> ('key, 'data) t
   -> Dyn.t
+
+(** {1 Deprecated}
+
+    The following is deprecated. Please migrate, and do not use in new code. *)
+
+(** This was renamed [set]. Hint: Run [ocamlmig migrate]. *)
+val replace : ('a, 'b) t -> key:'a -> data:'b -> unit
+[@@ocaml.deprecated "[since 2026-08] Use [Hashtbl.set]. Hint: Run [ocamlmig migrate]"]
+[@@migrate { repl = Rel.set }]
+
+(** This was renamed [set_seq]. Hint: Run [ocamlmig migrate]. *)
+val replace_seq : ('a, 'b) t -> ('a * 'b) Seq.t -> unit
+[@@ocaml.deprecated "[since 2026-08] Use [Hashtbl.set_seq]. Hint: Run [ocamlmig migrate]"]
+[@@migrate { repl = Rel.set_seq }]